requirements.txt
# Django
Django==4.1.5
djangorestframework==3.14.0
django-filter==22.1

# Clean code
black==22.12.0
isort==5.11.4

# Test utils
factory-boy==3.2.1
coverage==7.0.4
django-nose==1.4.7